ReducerCommand: Rework using core Reducer API
